% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Bal_XeeXhC}
\alias{Bal_XeeXhC}
\title{Calculate the correlation matrix for a balanced panel.}
\usage{
Bal_XeeXhC(dmat, X, e, n1, k)
}
\arguments{
\item{dmat}{the distance matrix}

\item{X}{the matrix of covariates}

\item{e}{the residuals}

\item{n1}{the number of observations}

\item{k}{the number of covariates}
}
\description{
Calculate the correlation matrix for a balanced panel.
}
