% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conley_spatial}
\alias{conley_spatial}
\title{Test dataset for spatial correlation.}
\format{A data frame with 5361 rows and 7 variables:
\describe{
  \item{EmpClean00}{Employment in 2000}
  \item{HDD}{Heating degree days}
  \item{CDD}{Cooling degree days}
  ...
}}
\source{
\url{https://github.com/darinchristensen/conley-se/tree/master/data}
}
\usage{
conley_spatial
}
\description{
A county~year panel dataset with employment and climate variables HDD and CDD
consisting of 362 counties and 15 years.
}
\keyword{datasets}
