% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley.R
\name{ConleySE.lm}
\alias{ConleySE.lm}
\title{Calculate spatially correlated variance-covariance matrix from a lm object.}
\usage{
\method{ConleySE}{lm}(model, x, y, kernel = "bartlett",
  dist_fn = "Haversine", dist_cutoff = 500, lat_scale = 111,
  verbose = FALSE)
}
\arguments{
\item{model}{a model object generated by \code{lm()}.}

\item{x}{a vector of x spatial coordinates.}

\item{y}{a vector of y spatial coordinates.}
}
\description{
Calculate spatially correlated variance-covariance matrix from a lm object.
}
