% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConleySE.R, R/conley.R
\docType{package}
\name{ConleySE}
\alias{ConleySE}
\alias{ConleySE-package}
\title{ConleySE-package}
\usage{
ConleySE(...)
}
\arguments{
\item{model}{a model object}

\item{kernel}{kernel for weights}

\item{dist_fn}{distance function}

\item{dist_cutoff}{distance for cutoff}

\item{lat_scale}{scale}

\item{verbose}{verbose output}
}
\value{
a variance-covariance matrix
}
\description{
Functions to calculate conley standard errors.

This function calculates standard errors for spatially correlated
data drawing from \href{https://doi.org/10.1016/S0304-4076(98)00084-0}{Conley (1999)}.
Primarily an adaptation of \href{https://github.com/darinchristensen/conley-se}{Darin Christensen}
with some changes for package consistency.
}
\author{
Nicholas Potter
}
