% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TimeDist}
\alias{TimeDist}
\title{Calculate the correlation matrix.}
\usage{
TimeDist(times, cutoff, X, e, n1, k)
}
\arguments{
\item{times}{vector of times}

\item{cutoff}{the cutoff distance}

\item{X}{the model matrix}

\item{e}{vector of residuals}

\item{n1}{number of obs}

\item{k}{number of variables}
}
\description{
Calculate the correlation matrix.
}
