% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley.R
\name{vcovConley.felm}
\alias{vcovConley.felm}
\title{Conley Spatial Variance-Covariances for Fixed Effects models fit with lfe.}
\usage{
\method{vcovConley}{felm}(model, id_var, time_var, x_var, y_var,
  kernel = "bartlett", dist_fn = "Haversine", dist_cutoff = 500,
  lag_cutoff = 5, lat_scale = 111, verbose = FALSE, cores = 1,
  balanced_pnl = FALSE)
}
\arguments{
\item{id_var}{character indicating individual variable}

\item{time_var}{character indicating time variable}

\item{x_var}{character name of variable of the x-axis (longitude)}

\item{y_var}{character name of variable of the y-axis (latitude)}

\item{lag_cutoff}{lag of cutoff}

\item{cores}{number of cores to use}

\item{balanced_pnl}{logical for whether balanced panel}
}
\description{
Conley Spatial Variance-Covariances for Fixed Effects models fit with lfe.
}
