% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley.R
\name{vcovConley.lm}
\alias{vcovConley.lm}
\title{Conley Spatial Variance-Covariances for linear model fits.}
\usage{
\method{vcovConley}{lm}(model, x, y, kernel = "bartlett",
  dist_fn = "Haversine", dist_cutoff = 500, lat_scale = 111,
  verbose = FALSE)
}
\arguments{
\item{model}{a model object generated by \code{lm()}.}

\item{x}{a vector of x spatial coordinates.}

\item{y}{a vector of y spatial coordinates.}
}
\description{
Conley Spatial Variance-Covariances for linear model fits.
}
