% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley.R
\name{vcovConley.matrix}
\alias{vcovConley.matrix}
\title{Conley Spatial Variance-Covariances for matrices.}
\usage{
\method{vcovConley}{matrix}(X, wgts, e, x, y, kernel, dist_fn, dist_cutoff,
  lat_scale, verbose)
}
\arguments{
\item{kernel}{kernel for weights.}

\item{dist_fn}{distance function.}

\item{dist_cutoff}{distance for cutoff.}

\item{lat_scale}{scale.}

\item{verbose}{verbose output.}

\item{model}{a model object.}
}
\value{
a variance-covariance matrix.
}
\description{
Conley Spatial Variance-Covariances for matrices.
}
