% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley.R
\name{vcovConley.plm}
\alias{vcovConley.plm}
\title{Conley Spatial Variance-Covariances for Panel Models fit with plm.}
\usage{
\method{vcovConley}{plm}(model, x, y, kernel = "bartlett",
  dist_fn = "Haversine", dist_cutoff = 500, lag_cutoff = 5,
  lat_scale = 111, verbose = FALSE, cores = 1,
  balanced_pnl = FALSE)
}
\description{
Conley Spatial Variance-Covariances for Panel Models fit with plm.
}
