% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley.R, R/vcovConley-package.R
\docType{package}
\name{vcovConley}
\alias{vcovConley}
\alias{vcovConley-package}
\title{Conley Spatially Correlated Variance-Covariances for Panel Models.}
\usage{
vcovConley(...)
}
\description{
This function calculates standard errors for spatially correlated
data drawing from \href{https://doi.org/10.1016/S0304-4076(98)00084-0}{Conley (1999)}.
Primarily an adaptation of \href{https://github.com/darinchristensen/conley-se}{Darin Christensen}
with some changes for package consistency.

Functions to calculate conley standard errors.
}
\seealso{
\code{\link{vcovConley.lm}}, \code{\link{vcovConley.plm}}, 
  \code{\link{vcovConley.felm}}, \code{\link{vcovConley.matrix}}
}
\author{
Nicholas Potter
}
