% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{XeeXhC_Lg}
\alias{XeeXhC_Lg}
\title{Calculate the correlation matrix for a large model. Avoids creating a 
distance matrix in memory but is slower.}
\usage{
XeeXhC_Lg(M, cutoff, X, e, n1, k, kernel = "bartlett",
  dist_fn = "Haversine")
}
\arguments{
\item{M}{a matrix of locations}

\item{cutoff}{the cutoff distance}

\item{X}{the model matrix}

\item{e}{vector of residuals}

\item{n1}{number of obs}

\item{k}{number of variables}

\item{kernel}{(string) a kernel function}

\item{dist_fn}{(string) a distance function}
}
\description{
Calculate the correlation matrix for a large model. Avoids creating a 
distance matrix in memory but is slower.
}
